move ball class into gametable
