+++ src/io/github/nrhoffmann/pong/gui/GameTable.java
@@ -1,5 +1,7 @@
 package io.github.nrhoffmann.pong.gui;
 
+import io.github.nrhoffmann.pong.physics.Vector;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseAdapter;
@@ -8,7 +10,7 @@ import java.awt.event.MouseEvent;
 class GameTable extends JPanel{
     private Ball ball = new Ball();
     private Timer timer;
-    public static final Dimension SIZE = new Dimension(1080, 720);
+    private static final Dimension SIZE = new Dimension(1080, 720);
 
     GameTable() {
         setPreferredSize(SIZE);
@@ -20,6 +22,13 @@ class GameTable extends JPanel{
             paintComponent(g);
         });
         timer.start();
+
+        addMouseMotionListener(new MouseAdapter() {
+            @Override
+            public void mouseMoved(MouseEvent e) {
+                System.out.println(e.getPoint());
+            }
+        });
     }
 
     @Override
@@ -31,7 +40,65 @@ class GameTable extends JPanel{
         Toolkit.getDefaultToolkit().sync();
     }
 
-    public Ball getBall() {
-        return new Ball();
+    public class Ball {
+        private Vector vector;
+        private Point location;
+        private int size;
+        private Color color;
+        private int maxY;
+        private int minY;
+
+        public void setColor(Color color) {
+            this.color = color;
+        }
+
+        public Color getColor() {
+            return color;
+        }
+
+        public void setLocation(Point location) {
+            this.location = location;
+        }
+
+        public Point getLocation() {
+            return location;
+        }
+
+        public void setSize(int size) {
+            this.size = size;
+        }
+
+        public int getSize() {
+            return size;
+        }
+
+        public void setVector(Vector vector) {
+            this.vector = vector;
+        }
+
+        public Vector getVector() {
+            return vector;
+        }
+
+        void tick() {
+            int deltax = (int) vector.getSpeedX();
+            int deltay = (int) vector.getSpeedY();
+            if (location.y + deltay < minY || location.y + deltay > maxY) {
+                deltay = Math.max(deltay, minY - location.y); //todo this might be ugly, or it might simulate mass compression
+                deltay = Math.min(deltay, maxY - location.y);
+                vector.bounceY();
+            }
+            location.translate(deltax, deltay);
+        }
+
+        Ball() {
+            size = 500;
+            color = new Color(128, 128, 128);
+            location = new Point(GameTable.SIZE.width / 2 - size / 2, GameTable.SIZE.height / 2);
+            vector = new Vector();
+
+            maxY = GameTable.SIZE.height - size;
+            minY = 0;
+        }
     }
 }
